Refactor: LinkCanisterModal
